Add files via upload
Define what orgs you don't want to query in the script. It then collects org names & org Ids, then uses the org ID to get network ID, then uses the network ID to get the L7 rules. It'll put all of that in an excel and then save it.

You must define your APIKey as an OS variable.